change n_val and n_test to int

+++ train.py
@@ -231,9 +231,9 @@ parser.add_argument('--n_val_tokens', type=int, default=-1, help='Number of uniq
                     in the validation dataset. If -1, then number tokens = (total - n_train_tokens) // 2.')
 parser.add_argument('--n_test_tokens', type=int, default=-1, help='Number of unique tokens to use \
                     in the test dataset. If -1, then number tokens = (total - n_train_tokens) // 2.')
-parser.add_argument('--n_val', type=float, default=-1,
+parser.add_argument('--n_val', type=int, default=-1,
                     help='Total # validation stimuli. Default: equal to n_train.')
-parser.add_argument('--n_test', type=float, default=-1,
+parser.add_argument('--n_test', type=int, default=-1,
                     help='Total # test stimuli. Default: equal to n_train.')
 parser.add_argument('--n_train_ood', nargs='+', required=False, default=[],
                     help='Size of OOD training sets.')
